refatcor: put corresponding commands to the dir.
