+++ command/meta/bles_meta_create.js
@@ -1,16 +1,17 @@
 const { Command, Argument } = require("commander");
 const chalk = require("chalk");
-const { WALLET_PATH } = require("../lib/constants");
+const anchor = require("@coral-xyz/anchor");
+const { WALLET_PATH } = require("../../lib/constants");
 const {
   getBlsContractClient,
   getPath,
   readKeypair,
   getMetadata,
   bs58Message,
-} = require("./utils");
+} = require("../utils");
 const { PublicKey } = require("@solana/web3.js");
 
-const blessMetaUpdateCommand = new Command("update")
+const blessMetaCreateCommand = new Command("create")
   .option(
     "--cluster <cluster>",
     "solana cluster: mainnet, testnet, devnet, localnet, <custom>",
@@ -22,7 +23,7 @@ const blessMetaUpdateCommand = new Command("update")
   .option(
     "--signer <signer>",
     "signer: the signer is the payer of the bless meta, default: " +
-      WALLET_PATH,
+    WALLET_PATH,
   )
   .option("--multisig <multisig>", "multisig:  the multisig of the bless meta")
   .option(
@@ -41,7 +42,7 @@ mint.required = true;
 const uri = new Argument("uri", "the meta data uri.");
 uri.required = true;
 
-blessMetaUpdateCommand
+blessMetaCreateCommand
   .addArgument(mint)
   .addArgument(uri)
   .action(async (mint, uri, options) => {
@@ -49,17 +50,23 @@ blessMetaUpdateCommand
     options.signer = options.signer || getPath(WALLET_PATH);
     options.squads = options.squads || false;
     try {
-      const metaJson = await getMetadata(uri);
       const keypair = readKeypair(options.signer);
       const client = getBlsContractClient(
         options.cluster,
         keypair,
         options.programId,
       );
+
+      const metaJson = await getMetadata(uri);
       let mintPubkey = new PublicKey(mint);
       const state =
         await client.blessTokenClient.getBlessTokenMetaState(mintPubkey);
       if (options.squads) {
+        if (options.multisig == null) {
+          console.log(chalk.red("multisig is required."));
+          process.exit(1);
+        }
+        const multisigPda = new PublicKey(options.multisig);
         if (options.admin == null) {
           console.log(chalk.red("admin is required."));
           process.exit(1);
@@ -68,13 +75,13 @@ blessMetaUpdateCommand
         if (state.admin.toBase58() != adminPubkey.toBase58()) {
           console.log(
             chalk.red(
-              "update is denied, admin is not matched, the state admin is " +
-                state.admin.toBase58(),
+              "create is denied, admin is not matched, the state admin is " +
+              state.admin.toBase58(),
             ),
           );
           process.exit(1);
         }
-        const tx = await client.blessTokenClient.getUpdateMetadataTx(
+        const tx = await client.blessTokenClient.getCreateMetadataTx(
           mintPubkey,
           adminPubkey,
           {
@@ -90,7 +97,7 @@ blessMetaUpdateCommand
           keypair,
         );
         console.log(
-          "bless token metadata account update transaction: \n" + itx,
+          "bless token metadata create account transaction created, " + itx,
         );
       } else {
         options.admin = options.admin || getPath(WALLET_PATH);
@@ -98,27 +105,32 @@ blessMetaUpdateCommand
         if (state.admin.toBase58() != adminKeypair.publicKey.toBase58()) {
           console.log(
             chalk.red(
-              "update is denied, admin is not matched, the state admin is " +
-                state.admin.toBase58(),
+              "create is denied, admin is not matched, the state admin is " +
+              state.admin.toBase58(),
             ),
           );
           process.exit(1);
         }
-        await client.blessTokenClient.updateMetadata(
+        await client.blessTokenClient.createMetadata(
           mintPubkey,
           adminKeypair.publicKey,
+          {
+            name: metaJson.name,
+            symbol: metaJson.symbol,
+            uri,
+          },
           {
             signer: keypair.publicKey,
             signerKeypair: [keypair, adminKeypair],
           },
         );
       }
-      console.log(chalk.green("bless token metadata account update success."));
+      console.log(chalk.green("bless token metadata create account success."));
       process.exit(0);
     } catch (e) {
-      console.log(chalk.red("bless token metadata account update fail: " + e));
+      console.log(chalk.red("bless token metadata create account fail: " + e));
       process.exit(1);
     }
   });
 
-module.exports = blessMetaUpdateCommand;
+module.exports = blessMetaCreateCommand;
