+++ command/meta/bles_meta_update.js
@@ -1,15 +1,16 @@
 const { Command, Argument } = require("commander");
 const chalk = require("chalk");
-const { WALLET_PATH } = require("../lib/constants");
+const { WALLET_PATH } = require("../../lib/constants");
 const {
   getBlsContractClient,
   getPath,
   readKeypair,
+  getMetadata,
   bs58Message,
-} = require("./utils");
+} = require("../utils");
 const { PublicKey } = require("@solana/web3.js");
 
-const blessMetaSetPendingAdminCommand = new Command("pending-admin")
+const blessMetaUpdateCommand = new Command("update")
   .option(
     "--cluster <cluster>",
     "solana cluster: mainnet, testnet, devnet, localnet, <custom>",
@@ -21,43 +22,40 @@ const blessMetaSetPendingAdminCommand = new Command("pending-admin")
   .option(
     "--signer <signer>",
     "signer: the signer is the payer of the bless meta, default: " +
-      WALLET_PATH,
+    WALLET_PATH,
   )
-  .option(
-    "--admin <admin>",
-    "admin: the admin of the bless meta, default: " + WALLET_PATH,
-  )
+  .option("--multisig <multisig>", "multisig:  the multisig of the bless meta")
   .option(
     "--squads <true/false>",
     "squads: if squads true, use squads to signature, default is false.",
   )
+  .option(
+    "--admin <admin>",
+    "admin: the admin of the bless meta, the admin will as  payer in squads mode ",
+  )
   .description(
-    "pending-admin: set the pending admin of the bless meta, the value is base58",
+    "create-meta: create meta account of the bless meta, the value is base58",
   );
 const mint = new Argument("mint", "mint: the public key of the mint token");
 mint.required = true;
-const pending = new Argument(
-  "pending-admin",
-  "pending-admin: the pending admin of the bless meta",
-);
-pending.required = true;
+const uri = new Argument("uri", "the meta data uri.");
+uri.required = true;
 
-blessMetaSetPendingAdminCommand
+blessMetaUpdateCommand
   .addArgument(mint)
-  .addArgument(pending)
-  .action(async (mint, pending, options) => {
+  .addArgument(uri)
+  .action(async (mint, uri, options) => {
     options.cluster = options.cluster || "localnet";
     options.signer = options.signer || getPath(WALLET_PATH);
-
     options.squads = options.squads || false;
     try {
+      const metaJson = await getMetadata(uri);
       const keypair = readKeypair(options.signer);
       const client = getBlsContractClient(
         options.cluster,
         keypair,
         options.programId,
       );
-      let pendingAdmin = new PublicKey(pending);
       let mintPubkey = new PublicKey(mint);
       const state =
         await client.blessTokenClient.getBlessTokenMetaState(mintPubkey);
@@ -70,16 +68,20 @@ blessMetaSetPendingAdminCommand
         if (state.admin.toBase58() != adminPubkey.toBase58()) {
           console.log(
             chalk.red(
-              "set pending admin is denied, admin is not matched, the state admin is " +
-                state.admin.toBase58(),
+              "update is denied, admin is not matched, the state admin is " +
+              state.admin.toBase58(),
             ),
           );
           process.exit(1);
         }
-        const tx = await client.blessTokenClient.getSetPendingAdminAccountTx(
+        const tx = await client.blessTokenClient.getUpdateMetadataTx(
           mintPubkey,
           adminPubkey,
-          pendingAdmin,
+          {
+            name: metaJson.name,
+            symbol: metaJson.symbol,
+            uri,
+          },
           { signer: adminPubkey },
         );
         const itx = await bs58Message(
@@ -88,7 +90,7 @@ blessMetaSetPendingAdminCommand
           keypair,
         );
         console.log(
-          "bless meta set pending admin transaction created: \n" + itx,
+          "bless token metadata account update transaction: \n" + itx,
         );
       } else {
         options.admin = options.admin || getPath(WALLET_PATH);
@@ -96,28 +98,27 @@ blessMetaSetPendingAdminCommand
         if (state.admin.toBase58() != adminKeypair.publicKey.toBase58()) {
           console.log(
             chalk.red(
-              "set pending admin is denied, admin is not matched, the state admin is " +
-                state.admin.toBase58(),
+              "update is denied, admin is not matched, the state admin is " +
+              state.admin.toBase58(),
             ),
           );
           process.exit(1);
         }
-        await client.blessTokenClient.setPendingAdminAccount(
+        await client.blessTokenClient.updateMetadata(
           mintPubkey,
           adminKeypair.publicKey,
-          pendingAdmin,
           {
             signer: keypair.publicKey,
             signerKeypair: [keypair, adminKeypair],
           },
         );
       }
-      console.log(chalk.green("bless meta set pending admin success."));
+      console.log(chalk.green("bless token metadata account update success."));
       process.exit(0);
     } catch (e) {
-      console.log(chalk.red("bless meta set the pending admin fail: " + e));
+      console.log(chalk.red("bless token metadata account update fail: " + e));
       process.exit(1);
     }
   });
 
-module.exports = blessMetaSetPendingAdminCommand;
+module.exports = blessMetaUpdateCommand;
