+++ command/program/index.js
@@ -1,0 +1,3 @@
+const programCommand = require("./program.js");
+
+module.exports = programCommand;
