+++ command/register/index.js
@@ -1,0 +1,3 @@
+const registerCommand = require("./register.js")
+
+module.exports = registerCommand;
