+++ command/register/reg_accept_admin.js
@@ -1,9 +1,8 @@
 const { Command, Argument } = require("commander");
 const anchor = require("@coral-xyz/anchor");
 const chalk = require("chalk");
-const { WALLET_PATH } = require("../lib/constants");
-const { getBlsRegisterClient, getPath, readKeypair } = require("./utils");
-const { PublicKey } = require("@solana/web3.js");
+const { WALLET_PATH } = require("../../lib/constants");
+const { getBlsRegisterClient, getPath, readKeypair } = require("../utils");
 
 const registerAcceptAdminCommand = new Command("accept-admin")
   .option(
