+++ command/register/reg_init.js
@@ -1,6 +1,6 @@
 const { Command, Argument } = require("commander");
-const { getBlsRegisterClient, getPath, readKeypair } = require("./utils");
-const { WALLET_PATH } = require("../lib/constants");
+const { getBlsRegisterClient, getPath, readKeypair } = require("../utils");
+const { WALLET_PATH } = require("../../lib/constants");
 const anchor = require("@coral-xyz/anchor");
 const chalk = require("chalk");
 const regInitCommand = new Command("init")
