+++ command/register/reg_pause.js
@@ -1,8 +1,7 @@
 const { Command, Argument } = require("commander");
-const anchor = require("@coral-xyz/anchor");
 const chalk = require("chalk");
-const { WALLET_PATH } = require("../lib/constants");
-const { getBlsRegisterClient, getPath, readKeypair } = require("./utils");
+const { WALLET_PATH } = require("../../lib/constants");
+const { getBlsRegisterClient, getPath, readKeypair } = require("../utils");
 
 const registerSetPauseCommand = new Command("pause")
   .option(
