+++ command/register/reg_set_deadline.js
@@ -1,8 +1,8 @@
 const { Command, Argument } = require("commander");
 const anchor = require("@coral-xyz/anchor");
 const chalk = require("chalk");
-const { getBlsRegisterClient, getPath, readKeypair } = require("./utils");
-const { WALLET_PATH } = require("../lib/constants");
+const { getBlsRegisterClient, getPath, readKeypair } = require("../utils");
+const { WALLET_PATH } = require("../../lib/constants");
 
 const registerSetDeadlineCommand = new Command("deadline")
   .option(
@@ -12,7 +12,7 @@ const registerSetDeadlineCommand = new Command("deadline")
   .option(
     "--signer <signer>",
     "signer: the signer is the backend signer who create the registraction account on chain. default " +
-      WALLET_PATH,
+    WALLET_PATH,
   )
   .description(
     "deadline: set/show the deadline of the registration, if is null, show daedline",
