+++ command/register/register.js
