+++ command/time/index.js
@@ -1,0 +1,3 @@
+const timeCommand = require("./time.js")
+
+module.exports = timeCommand;
