+++ command/time/time.js
