+++ command/time/time_set_pause.js
@@ -1,24 +1,21 @@
 const { Command, Argument } = require("commander");
-const anchor = require("@coral-xyz/anchor");
-const chalk = require("chalk");
-const { WALLET_PATH } = require("../lib/constants");
 const {
   getBlsTimeContractClient,
   getPath,
   readKeypair,
   bs58Message,
-} = require("./utils");
+} = require("../utils");
+const { WALLET_PATH } = require("../../lib/constants");
+const chalk = require("chalk");
 const { PublicKey } = require("@solana/web3.js");
-
-const timeSetPendingAdminCommand = new Command("pending-admin")
+const timeSetPausedCommand = new Command("set-paused")
   .option(
     "--cluster <cluster>",
     "solana cluster: mainnet, testnet, devnet, localnet, <custom>",
   )
   .option(
     "--signer <signer>",
-    "signer: the signer is the admin of the time contract, default: " +
-      WALLET_PATH,
+    "the signer is the admin of  the time contract: " + WALLET_PATH,
   )
   .option(
     "--programId <programId>",
@@ -32,32 +29,21 @@ const timeSetPendingAdminCommand = new Command("pending-admin")
     "--admin <admin>",
     "pending: if squads true, use admin to signature in squads",
   )
-  .description(
-    "pending-admin: set the pending admin of the registration, the value is base58, if is empty",
-  );
+  .description("merkle-root: set the merkle tree root.");
+const paused = new Argument("paused", "paused: set the time state paused.");
 const mint = new Argument(
   "mint",
   "mint: the mint is the mint token base58 value ",
 );
 mint.required = true;
-const pending = new Argument(
-  "pending-admin",
-  "pending-admin: the pending admin of the registration",
-);
-pending.required = true;
-timeSetPendingAdminCommand
+timeSetPausedCommand
   .addArgument(mint)
-  .addArgument(pending)
-  .action(async (mint, pending, options) => {
+  .addArgument(paused)
+  .action(async (mint, paused, options) => {
     options.cluster = options.cluster || "localnet";
     options.signer = options.signer || getPath(WALLET_PATH);
     try {
       const keypair = readKeypair(options.signer);
-      const client = getBlsTimeContractClient(
-        options.cluster,
-        keypair,
-        options.programId,
-      );
       let mintPubkey = null;
       try {
         mintPubkey = new PublicKey(mint);
@@ -65,29 +51,36 @@ timeSetPendingAdminCommand
         console.log(chalk.red("invaild mint parameter: " + e));
         process.exit(1);
       }
-
-      const state = await client.blessTimeClient.getTimeState(mintPubkey);
-      let pendingAdmin;
-      try {
-        pendingAdmin = new PublicKey(pending);
-      } catch (e) {
-        console.log(chalk.red("invalid pending-admin parameter: " + e));
+      const client = getBlsTimeContractClient(
+        options.cluster,
+        keypair,
+        options.programId,
+      );
+      if (paused != "true" && paused != "false") {
+        console.log(chalk.red("paused must be true or false."));
         process.exit(1);
       }
+      paused = paused == "true" ? true : false;
+      const state = await client.blessTimeClient.getTimeState(mintPubkey);
       if (options.squads) {
+        if (options.admin == null) {
+          console.log(chalk.red("admin is required."));
+          process.exit(1);
+        }
         const admin = new PublicKey(options.admin);
         if (state.adminAccount.toBase58() != admin.toBase58()) {
           console.log(
             chalk.red(
-              "set pending admin is denied, admin is not matched, the state  admin is " +
-                state.adminAccount.toBase58(),
+              "set paused is denied, admin is not matched, the state  admin is " +
+              state.adminAccount.toBase58(),
             ),
           );
           process.exit(1);
         }
-        const tx = await client.blessTimeClient.getSetPendingAdminAccountTx(
+
+        const tx = await client.blessTimeClient.getSetPausedTx(
           mintPubkey,
-          pendingAdmin,
+          paused,
           {
             signer: admin,
           },
@@ -97,30 +90,27 @@ timeSetPendingAdminCommand
           tx.instructions,
           keypair,
         );
-        console.log(
-          "bless time set pending admin transaction created: \n" + itx,
-        );
+        console.log("bless time set paused transaction created: \n" + itx);
       } else {
+        const keypair = readKeypair(options.signer);
         if (state.adminAccount.toBase58() != keypair.publicKey.toBase58()) {
           console.log(
             chalk.red(
-              "set pending admin is denied, admin is not matched, the state  admin is " +
-                state.adminAccount.toBase58(),
+              "set paused is denied, admin is not matched, the state  admin is " +
+              state.adminAccount.toBase58(),
             ),
           );
           process.exit(1);
         }
-        await client.blessTimeClient.setPendingAdminAccount(
-          mintPubkey,
-          pendingAdmin,
-        );
-        console.log(chalk.green("time contract set pending admin success."));
+
+        await client.blessTimeClient.setPaused(mintPubkey, paused);
+        console.log(chalk.green("set paused success."));
+        process.exit(0);
       }
-      process.exit(0);
     } catch (e) {
-      console.log(chalk.red("set the pending admin fail: " + e));
+      console.log(chalk.red("set paused fail: " + e));
       process.exit(1);
     }
   });
 
-module.exports = timeSetPendingAdminCommand;
+module.exports = timeSetPausedCommand;
