+++ command/tools/index.js
@@ -1,0 +1,3 @@
+const toolsCommand = require("./tools.js");
+
+module.exports = toolsCommand;
