+++ command/tools/tools.js
